Django/angela/yoga/requirement.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.13
distro==1.8.0
Django==4.1.7
django-environ==0.10.0
django-taggit==3.1.0
environ==1.0
idna==3.4
netaddr==0.8.0
Pillow==9.4.0
psycopg2-binary==2.9.6
requests==2.28.2
sqlparse==0.4.3
stripe==5.4.0
urllib3==1.26.15
wrapt==1.15.0
yookassa==2.3.5
